requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
faiss-cpu
python-dotenv
tiktoken
